// routes/logic/classes/interfaces/ProviderAddressInterface.ts
export interface ProviderAddressInterface {
    providerName: string;
    providerAddressCountry: string;
    providerAddressStreet: string;
    providerAddressPostcode: string;
    providerAddressCity: string;
    providerAddressURL: string;
    providerAddressTelefon: string;
    providerAddressFax: string;
    providerAddressEmail: string;
}

export interface ProviderAddressDataInterface {
    PROVIDERS_NAME: string;
    PROVIDERS_ADDR_COUNTRY: string;
    PROVIDERS_ADDR_STREET: string;
    PROVIDERS_ADDR_POSTCODE: string;
    PROVIDERS_ADDR_CITY: string;
    PROVIDERS_ADDR_URL: string;
    PROVIDERS_ADDR_TELEFON: string;
    PROVIDERS_ADDR_FAX: string;
    PROVIDERS_ADDR_EMAIL: string;
}